Allow setting of the maximum number of times to retry a Redis command
Add a configuration parameter, OIDCRedisCacheMaxRetry, to control the
maximum number of times that a Redis command should be retried. By
default the maximum retry count is set to 1, matching the current logic
but 0 can be used to disable retries or a larger value can be used to
retry more times.